requirements.txt
flask
torch
transformers
flask_cors
python-dotenv